pyproject.toml
[build-system]
requires = ["setuptools>=72"]
build-backend = "setuptools.build_meta"

[project]
name = "modelmetry-sdk"
version = "0.0.5"
description = "Modelmetry SDK for Python applications."
readme = "README.md"
license = { file = "LICENSE" }
authors = [{ name = "Modelemtry", email = "support@modelmetry.com" }]
keywords = [
  "llm",
  "monitoring",
  "guardrails",
  "tracing",
  "observability",
  "modelmetry",
]
requires-python = ">=3.0"
dependencies = [
  "urllib3>=1.25.3",
  "python-dateutil>=2.9.0",
  "typing-extensions>=4.12.2",
  "devtools>=0.12.2",
  "python-dotenv>=1.0.1",
  "pytest>=8.2.2",
  "httpx>=0.28.1",
  "attrs=>21.3.0",
  "python-dateutil=>2.8.0",
]
classifiers = [
  "Topic :: Software Development :: Libraries",
  "Topic :: Software Development :: Libraries :: Python Modules",
  "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

[project.urls]
GitHub = "https://github.com/modelmetry/modelmetry-sdk-python"
Homepage = "https://modelmetry.com"
